refactor(api): use rs256 for user jwt management
Separated user and client jwt services. For user one user RS256 to allow
third party's backend to verify the token on their own with the public
key. Implemented jwks endpoint to share the public key.

Files with keys "private.pem" and "public.pem" are expected to be in the
apps/mtas-api/ and are generated with the command described in README.md